Revert "chore: rename foxyBalancesApi -> foxyApi"
This reverts commit 1efcc95df7a1a66374321986a143f48bfb3fa6af.
